fix: protected route
