+++ src/pages/workouts/index.tsx
@@ -2,16 +2,15 @@ import { NextPage } from 'next';
 import Hero from '@components/Hero';
 import Stats from '@components/Workouts/Stats';
 import PreviousWorkouts from '@components/Workouts/PreviousWorkouts';
+import ProtectedRoute from '@components/ProtectedRoute';
 
 const WorkoutPage: NextPage = () => {
-  // TODO: Reroute if not logged in
-
   return (
-    <>
+    <ProtectedRoute>
       <Hero imgUrl="/workout-hero.jpg" title="Alla träningspass" />
       <Stats />
       <PreviousWorkouts />
-    </>
+    </ProtectedRoute>
   );
 };
